requirements.txt
fastapi
uvicorn
python-dotenv
requests
sentence-transformers
numpy
pydantic

